Fixed data product ROOT dictionary name so ROOT 6.38 automatically picks it up.  Got gaus_hit_finder to build with the latest-and-greatest phlex. Made gaus_hit_finder-FORM crossover jsonnet file run over 99 entries instead of just 5.

+++ gaus_hit_finder/find_hits_with_gaussians.hpp
@@ -44,7 +44,7 @@ namespace examples {
 
     std::vector<int> long_max_hits_vec;    ///<Maximum number hits on a really long pulse train
     std::vector<int> long_pulse_width_vec; ///<Sets width of hits used to describe long pulses
-    int max_multi_hit; ///<maximum hits for multi fit
+    size_t max_multi_hit; ///<maximum hits for multi fit
     int area_method;     ///<Type of area calculation
     std::vector<double>
       area_norms_vec;       ///<factors for converting area to same units as peak height
